scraper/pyproject.toml
[tool.poetry]
name = "scraper"
version = "0.1.0"
description = ""
authors = ["Matt-Hurd <matt-hurd@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
beautifulsoup4 = "^4.12.3"
selenium = "^4.23.1"
webdriver-manager = "^4.0.2"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.1"
google = "^3.0.0"
google-cloud-storage = "^2.18.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
